+++ eelbrain/testnd.py
@@ -1,5 +1,10 @@
 """Statistical tests for multidimensional data in :class:`NDVar` objects"""
 __test__ = False
 
-from ._stats.testnd import (configure, t_contrast_rel, corr, ttest_1samp,
-    ttest_ind, ttest_rel, anova)
+from ._stats.testnd import (
+    t_contrast_rel, corr, ttest_1samp, ttest_ind, ttest_rel, anova)
+
+
+def configure(*args, **kwargs):
+    raise RuntimeError("This function has been removed. Please use "
+                       "eelbrain.configure() instead")